Variable object and functions
Try to keep the global variable in one object. This way, you can better find them when you need to edit the code.

And try, if possible, to make the function cleaner. One function gets data and send it to the next. the next function does something and send is to the next and the last function renders it. (functional programming)